# internal/sqlc/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.14.0

package sqlc

import (
	"database/sql"
)

type User struct {
	ID        int32
	Username  string
	Email     string
	Password  string
	Bio       sql.NullString
	Image     sql.NullString
	CreatedAt sql.NullTime
	UpdatedAt sql.NullTime
}